Change the runtime rules to use a TIMEOUT to turned of electronic dog if we come back to the house after being out for 10 minutes.
Add support for an "All Off" and a "Lights All On" command.
Document new commands (All Off and Lights All On).
Add missing break on Timeout related generation.
Change name of enum "GetTimeoutState" to "PushTimeoutState".
Add support for Timeout with no action.
Correct timeout test to allow for the end of time.